Move OpenEO modules under features/layers/openeo
