+++ packages/base/src/features/layers/openeo/templates.ts
