+++ packages/base/src/features/layers/openeo/validation.ts
@@ -1,6 +1,6 @@
 import { ProcessGraph, ProcessRegistry } from '@openeo/js-processgraphs';
 
-import { connect, IOpenEOConnectionInfo } from '../mainview/OpenEOTileLayer';
+import { connect, IOpenEOConnectionInfo } from './OpenEOTileLayer';
 
 export interface IValidationError {
   code?: string;
